4.20 + 4.21 blog delete needs to be authenticated

+++ part4/blog/controllers/blogs.js
@@ -9,17 +9,18 @@ blogsRouter.get('/', async (request, response) => {
 })
 
 blogsRouter.post('/', async (request, response) => {
-  if (request.token === null) {
-    return response.status(401).json({
-      'error': 'token missing or invalid'
-    })
-  }
-  const user = await User.findById(request.token.id)
-  if (user === null) {
-    return response.status(401).json({
-      'error': 'user no longer exists'
-    })
-  }
+  // if (request.token === null) {
+  //   return response.status(401).json({
+  //     'error': 'token missing or invalid'
+  //   })
+  // }
+  // const user = await User.findById(request.token.id)
+  // if (user === null) {
+  //   return response.status(401).json({
+  //     'error': 'user no longer exists'
+  //   })
+  // }
+  const user = await getUser(request, response)
   const blogData = { ... request.body, 'user': user._id }
   const blog = new Blog(blogData)
   const result = await blog.save()
@@ -42,9 +43,31 @@ blogsRouter.put('/:id', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const user = await getUser(request, response)
   const id = request.params.id
+  const blog = await Blog.findById(id)
+  if (user.id !== blog.user.toString()) {
+    return response.status(401).json({ 'error': 'not authorized' })
+  }
   await Blog.findByIdAndRemove(id)
   return response.status(204).end()
 })
+
+const getUser = async (request, response) => {
+  if (request.token === null) {
+    response.status(401).json({
+      'error': 'token missing or invalid'
+    })
+    return
+  }
+  const user = await User.findById(request.token.id)
+  if (user === null) {
+    response.status(401).json({
+      'error': 'user no longer exists'
+    })
+    return
+  }
+  return user
+}
 
 module.exports = blogsRouter
